// src/queue/generated/handlers/IMessagesHandler.ts
/*
 * Copyright (c) Microsoft Corporation. All rights reserved.
 * Licensed under the MIT License. See License.txt in the project root for
 * license information.
 *
 * Code generated by Microsoft (R) AutoRest Code Generator.
 * Changes may cause incorrect behavior and will be lost if the code is
 * regenerated.
 */
// tslint:disable:max-line-length

import * as Models from "../artifacts/models";
import Context from "../Context";

export default interface IMessagesHandler {
  dequeue(options: Models.MessagesDequeueOptionalParams, context: Context): Promise<Models.MessagesDequeueResponse>;
  clear(options: Models.MessagesClearOptionalParams, context: Context): Promise<Models.MessagesClearResponse>;
  enqueue(queueMessage: Models.QueueMessage, options: Models.MessagesEnqueueOptionalParams, context: Context): Promise<Models.MessagesEnqueueResponse>;
  peek(options: Models.MessagesPeekOptionalParams, context: Context): Promise<Models.MessagesPeekResponse>;
}